webhook/src/requirements.txt
attrs==23.2.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
prometheus-client==0.19.0
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.17.1
setuptools==69.0.3
websocket-client==1.7.0
wheel==0.42.0